requirement.txt
yfinance==0.2.36
pandas==2.1.0
pyyaml==6.0.1
psycopg2-binary==2.9.9
plotly==5.3.1